Cleanup and Documentation
Errors in the sandbox are now reported using notify-send. Alongside capturing STDERR. This fixes an issues where sandbox errors fail silently when launched from the DE.
Ensure Create and Edit respect the Default keyword
Make SECCOMP a privileged operation.
Handle::wait_for_signal now accepts a Duration directly. All uses have changed from 1 second to 100 milliseconds.
Updated documentation

+++ src/cli/create.rs
@@ -1,5 +1,8 @@
 //! Create a new profile.
-use crate::aux::{env::AT_HOME, profile::Profile};
+use crate::{
+    aux::{env::AT_HOME, profile::Profile},
+    cli::default,
+};
 use anyhow::Result;
 use std::fs::{self, File};
 
@@ -14,21 +17,26 @@ pub struct Args {
 }
 impl super::Run for Args {
     fn run(self) -> Result<()> {
-        let path = {
-            let path = Profile::user_profile(&self.profile);
-            if !path.exists() && !self.blank {
-                fs::copy(AT_HOME.join("config").join("new.toml"), &path)?;
+        if self.profile == "default" {
+            let args = default::Args::default();
+            args.run()
+        } else {
+            let path = {
+                let path = Profile::user_profile(&self.profile);
+                if !path.exists() && !self.blank {
+                    fs::copy(AT_HOME.join("config").join("new.toml"), &path)?;
+                }
+                path
+            };
+
+            if !path.exists() {
+                File::create(&path)?;
             }
-            path
-        };
 
-        if !path.exists() {
-            File::create(&path)?;
+            if Profile::edit(&path).is_err() {
+                fs::remove_file(&path)?;
+            }
+            Ok(())
         }
-
-        if Profile::edit(&path).is_err() {
-            fs::remove_file(&path)?;
-        }
-        Ok(())
     }
 }
